pyproject.toml
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
]
